Migrated custom logger to 'logger_functions' to avoid circular import errors; implemented 'pathlib.Path' compatibility for 'cpf.XRD_FitPattern.initiate()'
