+++ src/cpf/data_preprocess.py
@@ -3,22 +3,22 @@
 """
 About
 =====
-Provides (or will provide) a number of options for smmoothing and noise reduction 
+Provides (or will provide) a number of options for smmoothing and noise reduction
 of the images before processing.
 
 This applies filters to the whole image
 For example:
-    
+
     data_prepare = {"smooth": {"Gaussian": 2},
                     "cosmics": {},
                     "mask": Calib_mask,
                     "order": {"smooth", "mask"}}
 
 
-Current useage of intensity limits:    
+Current useage of intensity limits:
     "cosmics": {**see cosmics for structure**}
     "smooth": {"smooth": {"filter": "median", "kernel": 10}}
- 
+
 Returns: data_class
 
 
@@ -26,7 +26,7 @@ Implemented:
     - cosmic removal (via another script).
     - median filtering.
     - rolling ball background removal
-        
+
 Possible:
     - differernce of gaussians x-ray diffraction
 
@@ -40,21 +40,22 @@ import cpf.settings as settings
 import numpy as np
 import matplotlib.pyplot as plt
 import scipy as sp
-from cpf.XRD_FitPattern import logger
+# from cpf.XRD_FitPattern import logger
+from cpf.logger_functions import logger
 
 def image_adjust(data, image_process):
     """
     Performs image cleaning processes on diffraction images.
-    
+
 
     Parameters
     ----------
     data : data class
         Diffraction data to be cleaned.
     image_process : String, list, dictionary or settings class
-        String is cleaning process to be performed; list of processes or 
-        dictionary containing settings for the cleaning. The setting class must contain 
-        the property 'data_prepare'. 
+        String is cleaning process to be performed; list of processes or
+        dictionary containing settings for the cleaning. The setting class must contain
+        the property 'data_prepare'.
         Examples of licit values are below.
 
     Raises
@@ -66,7 +67,7 @@ def image_adjust(data, image_process):
     -------
     data : data class
         data class containing the cleaned data.
-        
+
     Examples of acceptable inputs:
     image_process = 'cosmics'
     image_process = ['cosmics', 'smooth', 'background']
@@ -76,24 +77,24 @@ def image_adjust(data, image_process):
                      "order": ["cosmics", "smooth"]}
                      }
     image_process.data_prepare = ...
-    
+
     If image_process is a dictionary, each sub-dictionary contains the settings for the data cleaning function.
-    See each function for full details of the possible variables. 
-    
+    See each function for full details of the possible variables.
+
     The image cleaning is performed in the order of the list or dictionary, and in the order of the entires unless
     the dictionary contains the option "order".
-    
-    The data cleaning only works on image plate data. An error will be created if 
-    it is called for non-imageplate X-ray data. 
-    
+
+    The data cleaning only works on image plate data. An error will be created if
+    it is called for non-imageplate X-ray data.
+
     """
 
     # issue error if the data type is not correct.
     if data.DispersionType != "AngleDispersive":
         err_str = f"The data cleaning only works on image plate data. This data is {data.DispersionType}"
-        raise ValueError(err_str)        
+        raise ValueError(err_str)
 
-    
+
     if isinstance(image_process, dict):
         if "order" in image_process:
             processes = image_process["order"]
@@ -103,15 +104,15 @@ def image_adjust(data, image_process):
         processes = [image_process]
     else:
         processes = image_process
-        
+
     for i in range(len(processes)):
         if isinstance(image_process, dict) and processes[i] in image_process:
             options =  image_process[processes[i]]
-        else: 
-            options=None           
+        else:
+            options=None
 
         if processes[i].lower() == "cosmics":
-            data = remove_cosmics(data, options=options) 
+            data = remove_cosmics(data, options=options)
         elif processes[i].lower() == "smooth":
             data = smooth_image(data, options=options)
         elif processes[i].lower() == "background":
@@ -121,10 +122,10 @@ def image_adjust(data, image_process):
         else:
             err_str = "Unknown image pre-process type: %s", processes[i]
             raise ValueError(err_str)
-            
+
     return data
-        
-        
+
+
 
 def remove_cosmics(data, options=None, settings_for_fit=None):
     """
@@ -143,7 +144,7 @@ def remove_cosmics(data, options=None, settings_for_fit=None):
         DESCRIPTION.
 
     """
-    
+
     # set defaults
     gain = 2.2
     sigclip = 3.0 # 3 is dioptas default
@@ -159,7 +160,7 @@ def remove_cosmics(data, options=None, settings_for_fit=None):
             params["objlim"] = options["objlim"]
         if "sigfrac" in options:
             params["sigfrac"] = options["sigfrac"]
-        # FIX ME: use argparse or someway of passing any argument into cosmics.  
+        # FIX ME: use argparse or someway of passing any argument into cosmics.
     elif settings_for_fit and "cosmics" in settings_for_fit.data_prepare.keys():
         if "gain" in settings_for_fit.data_prepare["cosmics"]:
             params["gain"] = settings_for_fit.data_prepare["cosmics"]["gain"]
@@ -172,7 +173,7 @@ def remove_cosmics(data, options=None, settings_for_fit=None):
         # FIX ME: use argparse or someway of passing any argument into cosmics.
     else: # options == None and settings_for_fit==None:
         pass
-    
+
     test = cosmicsimage(
         data.intensity,
         **params,
@@ -195,7 +196,7 @@ def remove_cosmics(data, options=None, settings_for_fit=None):
 
 def smooth_image(data, options=None, settings_for_fit=None):
     """
-    Smooth  the diffraction image with either a median or Gaussian filter. 
+    Smooth  the diffraction image with either a median or Gaussian filter.
 
     Parameters
     ----------
@@ -222,9 +223,9 @@ def smooth_image(data, options=None, settings_for_fit=None):
         prep = settings_for_fit
     else:
         prep = {"smooth": {"filter": "median", "kernel": 5}}
-    
+
     dt = data.intensity
-    
+
     if prep['smooth']['filter'].lower() == "median":
         med_filt = morphology.disk(prep['smooth']['kernel'])
         #dt = filters.median(dt, med_filt)
@@ -244,21 +245,21 @@ def smooth_image(data, options=None, settings_for_fit=None):
                  "'" +prep['smooth']['filter'] + "' is not a numpy function that can be used as a smoothing function."
              )
             raise ValueError(err_str)
-        
+
         # if "nan" in prep['smooth']['filter']:
         #     dt.data[dt.mask==True] = np.nan
-        #     print(dt.data[0])       
+        #     print(dt.data[0])
     else:
         err_str = (
              "Unknown image smoothing type."
          )
         raise ValueError(err_str)
-       
+
     # copy data back into data class
     msk = data.intensity.mask
     dt = np.ma.array(dt, mask=msk)
     data.intensity = dt
-            
+
     return data
 
 
@@ -266,27 +267,27 @@ def smooth_image(data, options=None, settings_for_fit=None):
 
 def rolling_ball_background(data, options=None, settings_for_fit=None):
     """
-    rolling ball background removal is discussed in He's text book on 2D XRD.   
+    rolling ball background removal is discussed in He's text book on 2D XRD.
 
-    It is recomended that the image is smoothed before rolling ball is run inorder to 
-    negate the effects of any dark pixels in the diffraction image. 
+    It is recomended that the image is smoothed before rolling ball is run inorder to
+    negate the effects of any dark pixels in the diffraction image.
 
     Parameters
     ----------
     data : data class
         Diffraction data to be cleaned.
     options : dictionary, optional
-        Contains options for background removal. 
+        Contains options for background removal.
     settings_for_fit : settings class, optional
-        Contains options for background removal. 
-        
+        Contains options for background removal.
+
     Returns
     -------
     data : data_class
         Data class with the updated data.
 
     """
-    
+
     if isinstance(options, dict):
         if "background" in options.keys():
             prep = options
@@ -298,77 +299,77 @@ def rolling_ball_background(data, options=None, settings_for_fit=None):
     elif isinstance(settings_for_fit, dict):
         prep = settings_for_fit
     else:
-        prep = {"background": {"kernel": 50}}    
-        
+        prep = {"background": {"kernel": 50}}
+
     original_data = data.duplicate()
-    
+
     if "smooth" in prep["background"]:
-        # FIXME: we need a median filter that accounts for the mask. 
-        # This is either a filter than dignors the mask or we make all the masked values nan and run a nanmedial filter. 
+        # FIXME: we need a median filter that accounts for the mask.
+        # This is either a filter than dignors the mask or we make all the masked values nan and run a nanmedial filter.
         if prep["background"]["smooth"] == "":
             prep["background"]["smooth"] = {"filter": "median", "kernel": 5}
         data = smooth_image(data, prep["background"]["smooth"])
     elif "smooth" in prep:
         if prep["smooth"] == "":
-            prep["smooth"] = {"filter": "median", "kernel": 5}        
+            prep["smooth"] = {"filter": "median", "kernel": 5}
         data = smooth_image(data, prep["smooth"])
-        
+
     dt = data.intensity
     dt.data[dt.mask==True] = np.max(dt)
-    #FIX ME: the removal here is a metian for no good reason. I am trying to make sure the maksed areas do not bleed back into the iamge proper 
+    #FIX ME: the removal here is a metian for no good reason. I am trying to make sure the maksed areas do not bleed back into the iamge proper
     # I do not know if it is necessary though.
-    
+
     def plot_result(image, background, sigma=None ):
         fig, ax = plt.subplots(nrows=1, ncols=3)
-    
+
         a = ax[0].imshow(image, cmap='jet')
         ax[0].set_title('Original image')
         ax[0].axis('off')
         #plt.colorbar(a,cax=ax[0])
         #ax[1].colorbar()
         fig.colorbar(a, ax=ax[0], orientation='horizontal')
-    
+
         b=ax[1].imshow(background, cmap='jet')
         ax[1].set_title(r'Background, radius=%i' % sigma)
         ax[1].axis('off')
         #plt.colorbar(b,cax=ax[1])
         #ax[1].colorbar()
         fig.colorbar(b, ax=ax[1], orientation='horizontal')
-    
+
         c=ax[2].imshow(image - background, cmap='jet')
         ax[2].set_title('Result')
         ax[2].axis('off')
         #plt.colorbar(c,cax=ax[2])
         #fig.colorbar#add_colorbar(a)
-    
+
         fig.colorbar(c, ax=ax[2], orientation='horizontal')
         fig.tight_layout()
-    
-    
+
+
     background = restoration.rolling_ball(dt, radius=prep['background']['kernel'])
-    
+
     plot_result(original_data.intensity, background, prep['background']['kernel'])
     plt.show()
 
     no_bg = original_data.intensity - background
-    
+
     # copy data back into data class
     msk = no_bg.mask
     no_bg = np.ma.array(no_bg, mask=msk)
     data.intensity = no_bg
-        
+
     return data
 
-    
-    
+
+
 def scale_by_background(data, options=None, settings_for_fit=None):
     """
-    Scales the intsinsities to that of a rolling ball background performed on a smoothed image. 
-    
-    
-    # assume that the backgound is proportional to the transmissivity of the anvils. 
+    Scales the intsinsities to that of a rolling ball background performed on a smoothed image.
+
+
+    # assume that the backgound is proportional to the transmissivity of the anvils.
     # Scale the intensities by the background and multiply back up.
-    
+
 
     Parameters
     ----------
@@ -384,7 +385,7 @@ def scale_by_background(data, options=None, settings_for_fit=None):
     None.
 
     """
-    
+
     if isinstance(options, dict):
         if "background" in options.keys():
             prep = options
@@ -396,17 +397,17 @@ def scale_by_background(data, options=None, settings_for_fit=None):
     elif isinstance(settings_for_fit, dict):
         prep = settings_for_fit
     else:
-        prep = {"background": {"kernel": 200, "smooth": {"filter": "median", "kernel": 5}}} 
+        prep = {"background": {"kernel": 200, "smooth": {"filter": "median", "kernel": 5}}}
 
     original_data = data.duplicate()
-    
+
     background = rolling_ball_background(data, prep)
-    
+
     no_bg = original_data.intensity / (original_data.intensity-background.intensity)
-    
+
     # copy data back into data class
     msk = no_bg.mask
     no_bg = np.ma.array(no_bg, mask=msk)
     data.intensity = no_bg
-        
+
     return data
