+++ src/cpf/input_types/DioptasFunctions.py
@@ -20,7 +20,8 @@ from cpf.input_types._AngleDispersive_common import _AngleDispersive_common
 from cpf.input_types._Masks import _masks
 from cpf import IO_functions
 import cpf.h5_functions as h5_functions
-from cpf.XRD_FitPattern import logger
+# from cpf.XRD_FitPattern import logger
+from cpf.logger_functions import logger
 import cpf.logger_functions as lg
 
 
@@ -51,13 +52,13 @@ class DioptasDetector:
         self.tth_end   = None
         self.DispersionType = "AngleDispersive"
         self.continuous_azm = True
-        
+
         self.azm_blocks = 45
-        
+
         self.calibration = None
         self.conversion_constant = None
         self.detector = None
-        
+
         if settings_class:
             self.get_calibration(settings=settings_class)
         if self.calibration:
@@ -68,7 +69,7 @@ class DioptasDetector:
     def duplicate(self):
         """
         Makes an independent copy of a Dioptas Instance
-                
+
         Parameters
         ----------
         None.
@@ -84,12 +85,12 @@ class DioptasDetector:
 
     def get_calibration(self, file_name=None, settings=None):
         """
-        Opens the file with the calibration data in it and updates 
-        DioptasClass.calibration and 
+        Opens the file with the calibration data in it and updates
+        DioptasClass.calibration and
         DioptasClass.conversion_constant
-        
-        Either file_name or settings should be set. 
-        
+
+        Either file_name or settings should be set.
+
         Parameters
         ----------
         file_name : string, optional
@@ -118,13 +119,13 @@ class DioptasDetector:
 
     def get_detector(self, settings=None, calibration_file=None, diffraction_data=None, debug=False):
         """
-        Takes the detector information from the settings class, or the 
-        calibration *.json file and creates a detector-type instance which converts 
-        the x,y,azimith of the diffraction data pixels into two-theta vs. 
+        Takes the detector information from the settings class, or the
+        calibration *.json file and creates a detector-type instance which converts
+        the x,y,azimith of the diffraction data pixels into two-theta vs.
         azimuth.
-        
-        Either settings or the file_name are required. 
-        
+
+        Either settings or the file_name are required.
+
         Parameters
         ----------
         file_name : string, optional
@@ -144,11 +145,11 @@ class DioptasDetector:
         """
         if self.calibration == None:
             self.get_calibration(settings=settings, file_name=calibration_file, debug=debug)
-        
+
         if self.calibration:
             #use the poni file/calibration to make the detector
             self.detector = AzimuthalIntegrator(
-                detector=self.calibration.detector, 
+                detector=self.calibration.detector,
                 dist=self.calibration.dist,
                 poni1=self.calibration.poni1,
                 poni2=self.calibration.poni2,
@@ -157,15 +158,15 @@ class DioptasDetector:
                 rot3=self.calibration.rot3,
                 wavelength=self.calibration.wavelength,
             )
-            
-            if (self.detector.detector.get_name() == "Detector" and 
+
+            if (self.detector.detector.get_name() == "Detector" and
                 "detector_config" in self.calibration.as_dict()):
-                
+
                 config = self.calibration.as_dict()["detector_config"]
-                
+
                 if config["max_shape"] == None:
-                    # open the file to get the shape of the data.     
-                    if diffraction_data is not None: 
+                    # open the file to get the shape of the data.
+                    if diffraction_data is not None:
                         im_all = fabio.open(diffraction_data)
                     elif settings.calibration_data is not None:
                         im_all = fabio.open(settings.calibration_data)
@@ -173,24 +174,24 @@ class DioptasDetector:
                         im_all = fabio.open(settings.image_list[0])
                     if im_all:
                         config["max_shape"] = im_all.shape
-                        
+
                 #make sure the pixel sizes are correct
                 self.detector.detector.set_config(config)
-                            
-        
-   
+
+
+
     # @staticmethod
     def import_image(self, image_name=None, settings=None, mask=None, dtype=None, debug=False):
         """
-        Import the data image into the intensity array.        
+        Import the data image into the intensity array.
         Apply new mask to the data (if given) otherwise use previous mask
-    
+
         Parameters
         ----------
         image_name : string, optional
             Name of the image set to import. Either this or settings are required.
         settings : settings class, optional
-            Cpf setting class that constins the image set to import. 
+            Cpf setting class that constins the image set to import.
             Either this or imagename is required.
         mask : array, optional
             Mask array to apply to data. The default is None.
@@ -198,17 +199,17 @@ class DioptasDetector:
             Data type string, to force the data type and bit depth. The default is None.
         debug : boolian, optional
             True/Flase to display debuging information. The default is False.
-    
+
         Raises
         ------
         ValueError
             If there is no image_name of the settings.subpattern is not set.
-    
+
         Returns
         -------
         Im : masked array
             Masked image intensity array.
-    
+
         """
         # FIX ME: nxs files need to be checked. But they should be read like h5 files.
 
@@ -222,7 +223,7 @@ class DioptasDetector:
         if image_name == None:
             #load the data for the chosen subpattern.
             image_name = settings.subfit_filename
-            
+
         #read image
         if isinstance(image_name, list):
             # then it is a h5 type file
@@ -244,10 +245,10 @@ class DioptasDetector:
         if dtype==None:
             if self.intensity is None and np.size(self.intensity) > 2:
                 # self.intensity has been set before. Inherit the dtype.
-                dtype = self.intensity.dtype  
+                dtype = self.intensity.dtype
             elif "int" in im[0].dtype.name:
                 #the type is either int or uint - convert to float
-                # using same bit precision 
+                # using same bit precision
                 precision = re.findall("\d+", im[0].dtype.name)[0]
                 dtype = np.dtype("float"+precision)
         im = ma.array(im, dtype=dtype)
@@ -277,22 +278,22 @@ class DioptasDetector:
             self.intensity = ma.array(im, mask=self.intensity.mask)
             return ma.array(im)
 
-        
+
 
     def fill_data(
-        self, diff_file=None, settings=None, mask=None, make_zyx=False, 
+        self, diff_file=None, settings=None, mask=None, make_zyx=False,
         debug=False
     ):
         """
         Initiates the data arrays.
-        Creates the intensity, two theta and azimuth arrays from the 
+        Creates the intensity, two theta and azimuth arrays from the
         Detector and the data.
-        
-        If diffraction data is provided this makes the intensity array otherwise 
+
+        If diffraction data is provided this makes the intensity array otherwise
         the intensity array is filled with zeros.
 
         It must be called after setting the detector but before import_data.
-        
+
         Parameters
         ----------
         diff_file : string, optional
@@ -301,11 +302,11 @@ class DioptasDetector:
             cpf settings class containing the calibration file name.
             The default is None.
         mask : string or dirctionarry, optional
-            Filename or dictionary of instructions to make data mask. 
+            Filename or dictionary of instructions to make data mask.
             The default is None.
         make_zyx : boolian, optional
             Switch to make x,y,z arrays for the pixels. These arrats are not
-            used by default. They exist incase ever needed. 
+            used by default. They exist incase ever needed.
             The default is False.
         debug : True/False, optional
             Additional output, used for debugging.
@@ -327,33 +328,33 @@ class DioptasDetector:
                 raise ValueError("Settings are given but no subpattern is set.")
         else:
             raise ValueError("No diffraction file or settings have been given.")
-            
+
         if mask==None:
             if settings.calibration_mask:# in settings.items():
                 mask = settings.calibration_mask
-        
+
         if self.detector == None:
             self.get_detector(settings=settings)
 
         #get the intensities (without mask)
         self.intensity = self.import_image(diff_file)
-        # FIXME: (June 2024) because of how self.detector is instanciated the 
-        # shape might not be correct (or recognised). Hence the check here and 
+        # FIXME: (June 2024) because of how self.detector is instanciated the
+        # shape might not be correct (or recognised). Hence the check here and
         # inclusion of the shape in the array getting.
 
         if tuple(self.intensity.shape) != tuple(self.detector.detector.max_shape):
             # cast both shapes to tuples to prevent list != tuple error.
             raise ValueError("The pixel size of the data and the detector are not the same")
-            
+
         self.tth = ma.array(np.rad2deg(self.detector.twoThetaArray(self.detector.detector.max_shape)))
         self.azm = ma.array(np.rad2deg(self.detector.chiArray(self.detector.detector.max_shape)))
         # self.dspace = self._get_d_space()
         if make_zyx:
-            zyx = self.detector.calc_pos_zyx() 
+            zyx = self.detector.calc_pos_zyx()
             self.z = zyx[0]
-            self.y = zyx[1] 
-            self.x = zyx[2] 
-        
+            self.y = zyx[1]
+            self.x = zyx[2]
+
         #get and apply mask
         mask_array = self.get_mask(mask, self.intensity)
         self.mask_apply(mask_array, debug=debug)
@@ -387,7 +388,7 @@ class DioptasDetector:
             )
         # FIX ME: check the detector type is valid.
         return detector
-    
+
 
 
     def get_requirements(self, parameter_settings=None):
@@ -444,7 +445,7 @@ class DioptasDetector:
         return required_list
 
 
-# add common function. 
+# add common function.
 DioptasDetector._get_d_space = _AngleDispersive_common._get_d_space
 DioptasDetector.conversion   = _AngleDispersive_common.conversion
 DioptasDetector.bins         = _AngleDispersive_common.bins
@@ -459,7 +460,7 @@ DioptasDetector.mask_restore = _masks.mask_restore
 DioptasDetector.mask_remove  = _masks.mask_remove
 
 #these methods are all called from _Plot_AngleDispersive as they are shared with other detector types.
-#Each of these methods remains here because they are called by higher-level functions: 
+#Each of these methods remains here because they are called by higher-level functions:
 DioptasDetector.plot_masked      = _Plot_AngleDispersive.plot_masked
 DioptasDetector.plot_fitted      = _Plot_AngleDispersive.plot_fitted
 DioptasDetector.plot_collected   = _Plot_AngleDispersive.plot_collected
