+++ src/cpf/input_types/_Plot_AngleDispersive.py
@@ -7,27 +7,28 @@ import numpy.ma as ma
 import matplotlib.pyplot as plt
 from matplotlib import gridspec, cm, colors, tri
 from cpf.histograms import histogram2d
-from cpf.XRD_FitPattern import logger
+# from cpf.XRD_FitPattern import logger
+from cpf.logger_functions import logger
 import cpf.logger_functions as lg
 
 
 class _Plot_AngleDispersive():
     """
-    This class cannot be imported as a stand alone class. Instead the methods 
+    This class cannot be imported as a stand alone class. Instead the methods
     contained within it are created as methods in the angle dispersive diffraction
-    classes (e.g. DioptasDetector and SERFlvpDetector).    
-    """    
-    
+    classes (e.g. DioptasDetector and SERFlvpDetector).
+    """
+
     def _dispersion_ticks(self, disp_ticks=None, unique=10, disp_lims = None):
         """
         Returns the labels for the dispersion axis/colour bars.
-    
+
         :param disp_ticks: -- unused for maintained for compatibility with MED functions
         :param unique:
         :param disp_lims:
         :return new_tick_positions:
         """
-                
+
         if disp_lims is not None:
             disp_lims = np.around(np.array(disp_lims) / self.azm_blocks) * self.azm_blocks
             disp_ticks = list(range(int(disp_lims[0]), int(disp_lims[1] + 1), int(self.azm_blocks)))
@@ -39,11 +40,11 @@ class _Plot_AngleDispersive():
             disp_ticks = list(range(int(disp_lims[0]), int(disp_lims[1] + 1), int(self.azm_blocks)))
         elif len(np.unique(self.azm)) <= unique:
             # N.B. len(np.unique(self.azm)) is very slow if the data set is large.
-            # as long as self.azm_start and self.azm_end are set we should never get here. 
+            # as long as self.azm_start and self.azm_end are set we should never get here.
             disp_ticks = np.unique(np.unique(self.azm))
         else: #len(np.unique(self.azm)) >= unique:
             # N.B. len(np.unique(self.azm)) is very slow if the data set is large.
-            # as long as self.azm_start and self.azm_end are set we should never get here. 
+            # as long as self.azm_start and self.azm_end are set we should never get here.
             disp_lims = np.array(
                 [np.min(self.azm.flatten()), np.max(self.azm.flatten())]
             )
@@ -51,20 +52,20 @@ class _Plot_AngleDispersive():
                 block = 45
             else:
                 block = self.azm_blocks
-            disp_lims = np.around(disp_lims / block) * block   
+            disp_lims = np.around(disp_lims / block) * block
             disp_ticks = list(range(int(disp_lims[0]), int(disp_lims[1] + 1), int(block)))
-            
+
         return disp_ticks
-    
-    
-    
+
+
+
     def plot_masked(self, fig_plot=None):
         """
         Plot all the information needed to mask the data well.
         :param fig:
         :return:
         """
-    
+
         x_plots = 3
         y_plots = 2
         spec = gridspec.GridSpec(
@@ -75,7 +76,7 @@ class _Plot_AngleDispersive():
             hspace=0.5,
             height_ratios=[2, 1],
         )
-    
+
         ax1 = fig_plot.add_subplot(spec[0])
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -103,7 +104,7 @@ class _Plot_AngleDispersive():
             limits=[0, 100],
         )
         ax3.set_title("Masked Data")
-    
+
         ax4 = fig_plot.add_subplot(spec[3])
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -113,17 +114,17 @@ class _Plot_AngleDispersive():
             y_axis="intensity",
             limits=[0, 100],
         )
-    
+
         ax5 = fig_plot.add_subplot(spec[4])
         # plot cdf of the intensities.
         # sort the data in ascending order
         x1 = np.sort(self.intensity.data)
         x2 = np.sort(self.intensity)
-    
+
         # get the cdf values of y
         y1 = np.arange(np.size(x1)) / float(np.size(x1))
         y2 = np.arange(np.size(x2)) / float(ma.count(x2))
-    
+
         # ax1 = fig_1.add_subplot(1, 1, 1)
         ax5.plot(
             x1,
@@ -134,7 +135,7 @@ class _Plot_AngleDispersive():
             y2,
         )
         ax5.set_title("CDF of the intensities")
-    
+
         ax6 = fig_plot.add_subplot(spec[5])
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -144,25 +145,25 @@ class _Plot_AngleDispersive():
             y_axis="intensity",
             limits=[0, 100],
         )
-        
-    
+
+
     def plot_range(self, fig_plot=None, range_bounds=[-np.inf, np.inf], azm_bounds=[-np.inf, np.inf]):
         """
-        Plot data within given range. 
+        Plot data within given range.
         add data to axes.
         :param ax:
         :param show:
         :return:
         """
-        
+
         self.set_limits(range_bounds=range_bounds, azm_bounds=range_bounds)
-    
+
         # match max and min of colour scales
         limits = {
             "max": self.intensity.max(),
             "min": self.intensity.min(),
         }
-    
+
         # plot data
         ax1 = fig_plot.add_subplot(1, 1, 1)
         self.plot_calibrated(
@@ -176,13 +177,13 @@ class _Plot_AngleDispersive():
         ax1.set_title("Data")
         locs, labels = plt.xticks()
         plt.setp(labels, rotation=90)
-           
+
         # tidy layout
         plt.tight_layout()
 
-          
-    
-    
+
+
+
     def plot_fitted(self, fig_plot=None, model=None, fit_centroid=None, plot_type="surface", orientation="horizontal"):
         """
         add data to axes.
@@ -190,15 +191,15 @@ class _Plot_AngleDispersive():
         :param show:
         :return:
         """
-        
+
         # match max and min of colour scales
         limits = {
             "max": np.max([self.intensity.max(), model.max()]),
             "min": np.min([self.intensity.min(), model.min()]),
         }
-    
+
         tight=False
-        
+
         if orientation=="horizontal":
             up = 3
             across =1
@@ -221,7 +222,7 @@ class _Plot_AngleDispersive():
             ax.append(axs[0])
             ax.append(axs[1])
             ax.append(axs[2])
-    
+
         # plot data
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -239,7 +240,7 @@ class _Plot_AngleDispersive():
             ax[0].set_title("Data")
         locs, labels = plt.xticks()
         # plt.setp(labels, rotation=90)
-        
+
         # plot model
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -264,7 +265,7 @@ class _Plot_AngleDispersive():
             ax[1].set_title("Model")
         locs, labels = plt.xticks()
         # plt.setp(labels, rotation=90)
-        
+
         # plot residuals
         self.plot_calibrated(
             fig_plot=fig_plot,
@@ -289,7 +290,7 @@ class _Plot_AngleDispersive():
             ax[2].set_title("Residuals")
         locs, labels = plt.xticks()
         # plt.setp(labels, rotation=90)
-    
+
         # organise the axes and labelling.
         if tight==True:
             if orientation == "horizontal":
@@ -308,7 +309,7 @@ class _Plot_AngleDispersive():
                 # fig_plot.rcParams['axes.titlepad'] = -14  # pad is in points...
                 ax[2].set_xlabel("Azimuth ($^\circ$)")
             else:
-                        
+
                 bottom0, top0 = ax[0].get_ylim()
                 bottom1, top1 = ax[1].get_ylim()
                 bottom2, top2 = ax[2].get_ylim()
@@ -327,26 +328,26 @@ class _Plot_AngleDispersive():
                 ax[0].xaxis.set_ticks_position('both')
                 ax[1].xaxis.set_ticks_position('both')
                 ax[2].xaxis.set_ticks_position('both')
-   
+
         # tidy layout
         plt.tight_layout()
-        
-        
-    
-    
+
+
+
+
     def plot_collected(
-        self, 
-        fig_plot=None, 
-        axis_plot=None, 
-        show="intensity", 
+        self,
+        fig_plot=None,
+        axis_plot=None,
+        show="intensity",
         colourmap="jet",
-        limits=[0, 99.9], 
+        limits=[0, 99.9],
         location='default',
         debug=False
     ):
         """
         Plots the collected data as images.
-    
+
         Parameters
         ----------
         fig_plot : TYPE, optional
@@ -361,31 +362,31 @@ class _Plot_AngleDispersive():
             DESCRIPTION. The default is [0, 99.9].
         location : TYPE, optional
             DESCRIPTION. The default is 'bottom'.
-    
+
         Returns
         -------
         TYPE
             DESCRIPTION.
-    
+
         """
-        
+
         if axis_plot==None and fig_plot == None:
             #make a figure
             fig_plot, axis_plot = plt.subplots()
             display=True
-        
+
         elif axis_plot == None:
             #figure exists
             #assume will be shown outside of this function
             raise ValueError("Axes is needed to plot the data into")
             # FIX ME: the figure exists and so we whould be ableto find the axes
-        
+
         elif fig_plot == None:
-            # axis exists but figure not refrenced. 
+            # axis exists but figure not refrenced.
             #assume will be shown outside of this function
             raise ValueError("A figure is needed to plot the data into")
             # FIX ME: the axes exists and so we whould be ableto find the figure
-        
+
         if isinstance(show, str) and show == "unmasked_intensity":
             plot_i = self.intensity.data
         elif isinstance(show, str) and show == "mask":
@@ -396,21 +397,21 @@ class _Plot_AngleDispersive():
             plot_i = show
         else:  # if show == "intensity"
             plot_i = self.intensity
-    
-        
+
+
         if np.ndim(plot_i) >2:
             # The parametrized images to be plotted
             def f(data, n):
                 return data[n]
         else:
-            # there is only 1 plane of data. 
+            # there is only 1 plane of data.
             # (single detector rather than multiple detectors or positions.)
             def f(data, n):
                 return data
-    
+
         IMax = np.percentile(plot_i, limits[1])
         IMin = np.percentile(plot_i, limits[0])
-        
+
         if limits[0] > 0 and limits[1] < 100:
             cb_extend = "both"
         elif limits[1] < 100:
@@ -419,16 +420,16 @@ class _Plot_AngleDispersive():
             cb_extend = "min"
         else:
             cb_extend = "neither"
-    
+
         im_num = 0
         the_plot = axis_plot.imshow(
             f(plot_i,im_num), vmin=IMin, vmax=IMax, cmap=colourmap
         )
-        
+
         axis_plot.set_xlabel("x")
         axis_plot.set_ylabel("y")
         axis_plot.invert_yaxis()
-    
+
         if location == "default":
             if plot_i.shape[-1] > plot_i.shape[-2]:
                 location = "bottom"
@@ -436,13 +437,13 @@ class _Plot_AngleDispersive():
             else:
                 location = "right"
                 fraction = 0.15
-    
+
         cb = fig_plot.colorbar(mappable=the_plot, ax=axis_plot, extend=cb_extend, fraction=fraction, location=location)
-    
+
         if np.ndim(plot_i) >2 and plot_i.shape[0]>1:
             # adjust the main plot to make room for the sliders
             fig_plot.subplots_adjust(bottom=0.25)
-            
+
             # Make a horizontal slider to control the image being viewed.
             axim = fig_plot.add_axes([0.25, 0.1, 0.65, 0.03])
             im_slider = plt.Slider(
@@ -452,22 +453,22 @@ class _Plot_AngleDispersive():
                 valmax=plot_i.shape[0],
                 valinit=0,
             )
-    
-            
+
+
             # The function to be called anytime a slider's value changes
             def update(val):
                 the_plot.imshow(f(plot_i, im_slider.val))
                 fig_plot.canvas.draw_idle()
             # FIXME: we should also be able tp update the title and the colour scale.
-            
+
             # register the update function with each slider
             im_slider.on_changed(update)
-        
+
         #if display==True:
         #    the_plot.show()
-    
-    
-    
+
+
+
     def plot_calibrated(
         self,
         fig_plot=None,
@@ -493,27 +494,27 @@ class _Plot_AngleDispersive():
         :param show:
         :return:
         """
-     
-        
+
+
         y_ticks = None
         x_ticks = None
-            
+
         if axis_plot==None and fig_plot == None:
             #make a figure
             fig_plot, axis_plot = plt.subplots()
-        
+
         elif axis_plot == None:
             #figure exists
             pass
-        
+
         elif fig_plot == None:
-            # axis exists but figure not refrenced. 
-            pass    
-        
-        if rastered==False and self.intensity.size > 1E6: #100000000000:# 1000000: # was 50000 until fixed max/min functions 
+            # axis exists but figure not refrenced.
+            pass
+
+        if rastered==False and self.intensity.size > 1E6: #100000000000:# 1000000: # was 50000 until fixed max/min functions
             logger.moreinfo(" ".join(map(str, [(" Have patience. The plot(s) will appear but it can take its time to render.")])))
             rastered = True
-    
+
         if x_axis == "azimuth":
             plot_x = self.azm
             x_lims = [self.azm_start, self.azm_end]
@@ -521,7 +522,7 @@ class _Plot_AngleDispersive():
             plot_x = self.tth
         # plot_y = self.azm
         label_x = x_label
-        
+
         if y_axis == "intensity":
             # plot y rather than azimuth on the y axis
             plot_y = self.intensity
@@ -538,9 +539,9 @@ class _Plot_AngleDispersive():
                 # y_lims = [self.azm.min(), self.azm.max()]
             # axis_plot.set_ylim(y_lims)
             # y_ticks = list(range(int(y_lims[0]),int(y_lims[1]+1),45))
-    
+
             y_ticks = self.dispersion_ticks()
-    
+
         # organise the data to plot
         if data is not None:
             plot_i = data
@@ -555,10 +556,10 @@ class _Plot_AngleDispersive():
             colourmap = "Greys"
         else:  # if show == "intensity"
             plot_i = plot_i
-    
+
         # set axis limits
         x_lims = [plot_x.min(), plot_x.max()]
-    
+
         # set colour bar and colour maps.
         if colourmap == "Greys":
             IMax = 2.01
@@ -607,7 +608,7 @@ class _Plot_AngleDispersive():
             location=="right"
         else:
             pass
-        
+
         # set colour map
         if colourmap == "residuals-blanaced":
             # colourmap = self.residuals_colour_scheme(
@@ -617,13 +618,13 @@ class _Plot_AngleDispersive():
         else:
             colourmap = colourmap
 
-        #if horizontal swap everything around. 
+        #if horizontal swap everything around.
         if orientation.lower() == "horizontal":
             plot_y, plot_x = plot_x, plot_y
             x_lims, y_lims = y_lims, x_lims
             label_x, label_y = label_y, label_x
-            
-            
+
+
             location = "right"
         else:
             location = "bottom"
@@ -666,23 +667,23 @@ class _Plot_AngleDispersive():
                 pixels_per_bin = point_scale,
                 resample_shape = resample_shape
             )
-            
+
         axis_plot.set_xlabel(label_x)
         axis_plot.set_ylabel(label_y)
-    
+
         axis_plot.set_xlim(x_lims)
         axis_plot.set_ylim(y_lims)
-        
+
         if y_axis!="intensity":
             if orientation=="horizontal":
                 axis_plot.set_xticks(y_ticks)
             else:
                 axis_plot.set_yticks(y_ticks)
-       
-            
-        #p2 = axis_plot.get_position().get_points().flatten()        
-        #ax_cbar1 = fig_plot.add_axes([p2[0], 0, p2[2]-p2[0], 0.05]) 
-    
+
+
+        #p2 = axis_plot.get_position().get_points().flatten()
+        #ax_cbar1 = fig_plot.add_axes([p2[0], 0, p2[2]-p2[0], 0.05])
+
         cb = fig_plot.colorbar(mappable=the_plot, ax=axis_plot, extend=cb_extend, location=location, shrink=0.9)#, pad=0.1, aspect=8)
 
 
@@ -725,9 +726,9 @@ def residuals_colour_scheme(maximum_value, minimum_value, **kwargs):
     )
 
     return all_colours
-   
-    
-    
+
+
+
 def raster_plot(
     # self,
     data_plot, x_plot, y_plot,
@@ -740,17 +741,17 @@ def raster_plot(
     pixels_per_bin=3,
 ):
     """
-    Converts the diffraction data into an image and plots it. It is a more efficient replacement for 
-    the default scatter plots. it should be called when the data sets are too large to plot in a 
-    timely manner. 
-    
+    Converts the diffraction data into an image and plots it. It is a more efficient replacement for
+    the default scatter plots. it should be called when the data sets are too large to plot in a
+    timely manner.
+
     In essence it integrates the diffraction data into a plot with the pixel resolution of the screen.
-    It calls the same numpy functions that pyFAI uses but without the same level of data cleaning. 
+    It calls the same numpy functions that pyFAI uses but without the same level of data cleaning.
 
     Parameters
     ----------
     data_plot : masked array
-        data to convert into the image to plot. 
+        data to convert into the image to plot.
     x_plot : masked array
         horizontal position of the data points (usually two theta)
     y_plot : masked array
@@ -775,21 +776,21 @@ def raster_plot(
     pl : axes
         filled set of axes.
     """
-    # FIX ME: for massive data sets (e.g. ESFR) this is still slow. 
-    # maybe it should be replaced by datashader or something else that is good for dynamicaly plotting massive data sets. 
-    # an alternative is to use GIS image referencing packages. 
-    
+    # FIX ME: for massive data sets (e.g. ESFR) this is still slow.
+    # maybe it should be replaced by datashader or something else that is good for dynamicaly plotting massive data sets.
+    # an alternative is to use GIS image referencing packages.
+
     if lg.make_logger_output(level="DEBUG"):
         import time
         start = time.time()
-    
-    
+
+
     if resample_shape != None:
         # use the size of the output set by resample_shape
         x_bins = resample_shape[0]
         y_bins = resample_shape[1]
     else:
-        #base the sixe of the output on the number of pixels in the axes. 
+        #base the sixe of the output on the number of pixels in the axes.
         if axis_plot == None:
             if fig_plot==None:
                 fig = plt.figure()
@@ -800,26 +801,26 @@ def raster_plot(
         height *= fig_plot.dpi
         x_bins = width/pixels_per_bin
         y_bins = height/pixels_per_bin
-        
+
     #FIX ME: should I just replace this with a call to histogram2d_engine in pyFAI/engines/histogram_engine.py??
-    # To have empyt pixels where there is no data use dummy = np.nan as an option. 
+    # To have empyt pixels where there is no data use dummy = np.nan as an option.
     # see comments in histogram2d for answer
     result, x_edges, y_edges, num_pix_per_bin  = histogram2d(data_plot, x_plot, y_plot,
                 x_bins = x_bins,
                 y_bins = y_bins)
     result = np.rot90(result)
     pl = axis_plot.imshow(result, extent=[x_edges[0],x_edges[-1],y_edges[0],y_edges[-1]], aspect="auto", vmin=vmin, vmax=vmax, cmap=colourmap)
-    
+
     #axis_plot.invert_yaxis()
     if lg.make_logger_output(level="DEBUG"):
         end = time.time()
         logger.debug(" ".join(map(str, [(f"Time to make rastered plot {end-start}")] )) )
-    
+
     return pl
 
 
 
-    
+
 def surface_plot(
     # self,
     data_plot, x_plot, y_plot,
@@ -837,7 +838,7 @@ def surface_plot(
     Parameters
     ----------
     data_plot : masked array
-        data to convert into the image to plot. 
+        data to convert into the image to plot.
     x_plot : masked array
         horizontal position of the data points (usually two theta)
     y_plot : masked array
@@ -862,13 +863,13 @@ def surface_plot(
     pl : axes
         filled set of axes.
     """
-    
+
     triang = tri.Triangulation(x_plot.flatten(), y_plot.flatten())
     corners = triang.triangles
     # FIXME: i might be better to make the triangles from the x,y, points rather than the tth+azm.
-    # when made like this the plots spread the intensity in a broadway. -- the scatter plot looks better. 
-    
-    
+    # when made like this the plots spread the intensity in a broadway. -- the scatter plot looks better.
+
+
     triang.set_mask(tri.TriAnalyzer(triang).get_flat_tri_mask())
     mask=np.zeros(len(triang.triangles))
     for i in range(len(corners)):
@@ -876,17 +877,17 @@ def surface_plot(
             mask[i]=True
     mask=np.array(mask, dtype="bool")
     triang.set_mask((triang.mask==True) | (mask==True))
-    
-    if 0:        
+
+    if 0:
         areas=[]
         x_range = []
         y_range = []
         import proglog
         progress_bar = proglog.default_bar_logger('bar')  # shorthand to generate a bar logger
-    
+
         def PolyArea(x,y):
             return 0.5*np.abs(np.dot(x,np.roll(y,1))-np.dot(y,np.roll(x,1)))
-        
+
         #for f in range(setting_class.image_number):
         for i in progress_bar.iter_bar(iteration=range(len(corners))):
             if triang.mask[i] == True:
@@ -897,15 +898,15 @@ def surface_plot(
                 y_range.append( y_plot.flatten()[triang.triangles][i].max() - y_plot.flatten()[triang.triangles][i].min() )
 
         fig, axs = plt.subplots(1,3, sharey=True, tight_layout=True)
-    
+
         # We can set the number of bins with the *bins* keyword argument.
         axs[0].hist(areas, bins=50)#int(len(areas)/40))
         axs[1].hist(x_range, bins=50)#int(len(areas)/40))
         axs[2].hist(y_range, bins=50)#int(len(areas)/40))
-    
-        # FIXME: I should filter the triangles on their area or their shape. 
+
+        # FIXME: I should filter the triangles on their area or their shape.
         # mean_area = np.mean(areas)
-    
+
     pl = axis_plot.tripcolor(triang, data_plot.flatten(),cmap=colourmap, vmin=vmin, vmax=vmax)
-    
+
     return pl
