+++ src/cpf/output_formatters/WriteFitMovie.py
@@ -11,7 +11,8 @@ import cpf.IO_functions as IO
 from cpf.XRD_FitSubpattern import plot_FitAndModel
 from cpf.data_preprocess import remove_cosmics as cosmicsimage_preprocess
 from cpf.BrightSpots import SpotProcess
-from cpf.XRD_FitPattern import logger
+# from cpf.XRD_FitPattern import logger
+from cpf.logger_functions import logger
 
 
 def Requirements():
@@ -30,10 +31,10 @@ def Requirements():
 
 def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
     """
-    Writes a *.?? file of the fits. 
-    
+    Writes a *.?? file of the fits.
+
     N.B. this output requires the data files to be present to work.
-    
+
     Parameters
     ----------
     FitSettings : TYPE
@@ -62,8 +63,8 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
         raise ValueError(
             "The frames per second needs to be a number."
         )
-        
-    
+
+
 
     if setting_class is None and setting_file is None:
         raise ValueError(
@@ -82,8 +83,8 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
     if base is None:
         logger.info(" ".join(map(str, [("No base filename, using input filename instead.")])))
         base = os.path.splitext(os.path.split(setting_class.settings_file)[1])[0]
-        
-    # make the data class.     
+
+    # make the data class.
     data_to_fill = setting_class.image_list[0]
     data_class = setting_class.data_class
     data_class.fill_data(
@@ -91,15 +92,15 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
         settings=setting_class,
         debug=debug,
     )
-    
-    
+
+
     duration = (setting_class.image_number)/fps
     num_subpatterns = len(setting_class.fit_orders)
-    
+
     for z in range(num_subpatterns):
-        
+
         y = list(range(setting_class.image_number))
-        
+
         setting_class.set_subpattern(0, z)
 
 
@@ -107,28 +108,28 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
         if setting_class.file_label != None:
             addd = addd + setting_class.file_label
         out_file = IO.make_outfile_name(
-            base, 
-            directory=setting_class.output_directory, 
-            extension=file_types[0], 
-            overwrite=True, 
+            base,
+            directory=setting_class.output_directory,
+            extension=file_types[0],
+            overwrite=True,
             additional_text=addd
         )
         logger.info(" ".join(map(str, [("Writing %s" % out_file)])))
-                    
+
         # this calls all the iamges and adds them as frames to the video.
         # edited after :https://zulko.github.io/moviepy/getting_started/working_with_matplotlib.html?highlight=matplotlib
         # 4th April 2023.
         def make_frame(t):
-        
-            # t scales between 0 and 1. 
-            # to call each of the images in turn t has to be scaled back 
-            # into the number of images (here 'y'). And it has to be an integer. 
+
+            # t scales between 0 and 1.
+            # to call each of the images in turn t has to be scaled back
+            # into the number of images (here 'y'). And it has to be an integer.
             #logger.info(" ".join(map(str, [(t, int(t*fps), y[int(t*fps)])])))
-                        
+
             # Get diffraction pattern to process.
             data_class.import_image(setting_class.image_list[y[int(t*fps)]], debug=debug)
-                      
-            
+
+
             if setting_class.datafile_preprocess is not None:
                 # needed because image preprocessing adds to the mask and is different for each image.
                 data_class.mask_restore()
@@ -137,8 +138,8 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
             else:
                # nothing is done here.
                pass
-                        
-            # restrict data to the right part.           
+
+            # restrict data to the right part.
             sub_data = data_class.duplicate()
             setting_class.set_subpattern(y[int(t*fps)], z)
             sub_data.set_limits(range_bounds=setting_class.subfit_orders["range"])
@@ -149,7 +150,7 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
                 or "imin" in setting_class.subfit_orders
             ):
                 sub_data = SpotProcess(sub_data, setting_class)
-            
+
             # read fit file
             json_file = IO.make_outfile_name(
                 setting_class.subfit_filename,  # diff_files[z],
@@ -159,26 +160,26 @@ def WriteOutput(setting_class=None, setting_file=None, debug=False, **kwargs):
             )
             with open(json_file) as json_data:
                 data_fit = json.load(json_data)[z]
-            
-            # make the plot of the fits. 
+
+            # make the plot of the fits.
             fig = plt.figure(1)
-            fig = plot_FitAndModel(setting_class, 
-                                   sub_data, 
-                                   #param_lmfit=None, 
-                                   params_dict=data_fit, 
+            fig = plot_FitAndModel(setting_class,
+                                   sub_data,
+                                   #param_lmfit=None,
+                                   params_dict=data_fit,
                                    figure=fig)
-            title_str = (IO.peak_string(setting_class.subfit_orders)+"; " + str(y[int(t*fps)]) +"/" + 
+            title_str = (IO.peak_string(setting_class.subfit_orders)+"; " + str(y[int(t*fps)]) +"/" +
                             str(setting_class.image_number) + "\n" +
                             IO.title_file_names(setting_class, num=y[int(t*fps)], image_name=setting_class.subfit_filename)
                             )
             if "note" in setting_class.subfit_orders:
                 title_str = title_str + " " + setting_class.subfit_orders["note"]
-            plt.suptitle(title_str)   
+            plt.suptitle(title_str)
             IO.figure_suptitle_space(fig, topmargin=0.4)
-            
-            #return the figure            
+
+            #return the figure
             return mplfig_to_npimage(fig)
-        
+
         # make the video clip
         animation = VideoClip(make_frame, duration=duration)
         animation.write_videofile(out_file, fps=fps)
