+++ src/cpf/peak_functions.py
@@ -3,7 +3,7 @@
 """
 Functions describing the peak shape and properties.
 
-Ideally if a new peak shape is needed it would be 
+Ideally if a new peak shape is needed it would be
 possible to define it here and it would propagate through the code.
 But this is not implemented and might not be possible...
 """
@@ -17,7 +17,8 @@ __all__ = [
 ]
 
 import numpy as np
-from cpf.XRD_FitPattern import logger
+# from cpf.XRD_FitPattern import logger
+from cpf.logger_functions import logger
 
 
 def peak_components(full=False, include_profile=True):
